refactor folder structure to make it align with clean architecture best practices alsorename variables and method names to make them follow best coding practices
