+++ backend/AAiT-backend-group-8/Delivery/Controller/user_controller.go
@@ -2,9 +2,8 @@ package controller
 
 import (
 	"AAiT-backend-group-8/Domain"
-	"net/http"
-
 	"github.com/gin-gonic/gin"
+	"net/http"
 )
 
 func (controller *Controller) RegisterUser(c *gin.Context) {
@@ -14,7 +13,7 @@ func (controller *Controller) RegisterUser(c *gin.Context) {
 		return
 	}
 
-	err := controller.UserUsecase.RegisterUser(&user)
+	err := controller.UserUseCase.RegisterUser(&user)
 	if err != nil {
 		if err.Error() == "email already exists" {
 			c.JSON(400, gin.H{"error": err.Error()})
@@ -34,7 +33,7 @@ func (controller *Controller) VerifyEmail(c *gin.Context) {
 		return
 	}
 
-	err := controller.UserUsecase.VerifyEmail(token)
+	err := controller.UserUseCase.VerifyEmail(token)
 	if err != nil {
 		c.JSON(500, gin.H{"error": err.Error()})
 		return
@@ -57,7 +56,7 @@ func (controller *Controller) Login(c *gin.Context) {
 		return
 	}
 
-	token, refresher, err := controller.UserUsecase.Login(ep.Email, ep.Password)
+	token, refresher, err := controller.UserUseCase.Login(ep.Email, ep.Password)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
@@ -69,13 +68,13 @@ func (controller *Controller) Login(c *gin.Context) {
 func (controller *Controller) RefreshToken(c *gin.Context) {
 	var cred Domain.Credential
 
-	bind_err := c.BindJSON(&cred)
-	if bind_err != nil {
+	bindErr := c.BindJSON(&cred)
+	if bindErr != nil {
 		c.IndentedJSON(400, gin.H{"message": "invalid request payload"})
 		return
 	}
 
-	token, err := controller.UserUsecase.RefreshToken(cred.Email, cred.Refresher)
+	token, err := controller.UserUseCase.RefreshToken(cred.Email, cred.Refresher)
 	if err != nil {
 		c.IndentedJSON(400, gin.H{"message": "invalid refresh token "})
 	}
@@ -90,7 +89,7 @@ func (controller *Controller) ForgotPassword(c *gin.Context) {
 		return
 	}
 
-	err := controller.UserUsecase.GenerateResetPasswordToken(email)
+	err := controller.UserUseCase.GenerateResetPasswordToken(email)
 	if err != nil {
 		if err.Error() == "user not found" {
 			c.JSON(404, gin.H{"error": "User not found"})
@@ -110,7 +109,7 @@ func (controller *Controller) StoreToken(c *gin.Context) {
 		return
 	}
 
-	err := controller.UserUsecase.StoreToken(token)
+	err := controller.UserUseCase.StoreToken(token)
 	if err != nil {
 		c.JSON(500, gin.H{"error": err.Error()})
 		return
@@ -130,7 +129,7 @@ func (controller *Controller) ResetPassword(c *gin.Context) {
 		return
 	}
 
-	err := controller.UserUsecase.ResetPassword(payload.Token, payload.NewPassword)
+	err := controller.UserUseCase.ResetPassword(payload.Token, payload.NewPassword)
 	if err != nil {
 		if err.Error() == "invalid or expired token" || err.Error() == "invalid token payload" || err.Error() == "invalid or mismatched token" {
 			c.JSON(400, gin.H{"error": err.Error()})
