+++ backend/AAiT-backend-group-8/Infrastructure/search_params.go
@@ -1,4 +1,4 @@
-package Helper
+package infrastructure
 
 import (
 	"AAiT-backend-group-8/Domain"
